# docs/swagger.yaml
basePath: /
definitions:
  models.Subscription:
    properties:
      endDate:
        type: string
      id:
        type: integer
      price:
        type: integer
      serviceName:
        type: string
      startDate:
        type: string
      userID:
        type: string
    type: object
  models.TotalCostRequest:
    properties:
      period_end:
        type: string
      period_start:
        type: string
      service_name:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для управления подписками
  title: Subscriptions Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Возвращает список подписок с фильтрацией по user_id и service_name
      parameters:
      - description: ID пользователя
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Subscription'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создает новую подписку с указанными данными
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать новую подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаляет подписку по указанному ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Возвращает подписку по указанному ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить подписку по ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет данные подписки по ID
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/total-cost:
    post:
      consumes:
      - application/json
      description: Рассчитывает общую стоимость подписок за период с фильтрами
      parameters:
      - description: Запрос на расчет
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.TotalCostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Рассчитать общую стоимость подписок
      tags:
      - subscriptions
swagger: "2.0"